+++ Vungle/com/mopub/mobileads/VungleInterstitial.java
@@ -49,9 +49,9 @@ public class VungleInterstitial extends CustomEventInterstitial {
 
     @Override
     protected void loadInterstitial(Context context,
-            CustomEventInterstitialListener customEventInterstitialListener,
-            Map<String, Object> localExtras,
-            Map<String, String> serverExtras) {
+                                    CustomEventInterstitialListener customEventInterstitialListener,
+                                    Map<String, Object> localExtras,
+                                    Map<String, String> serverExtras) {
         mCustomEventInterstitialListener = customEventInterstitialListener;
         mIsPlaying = false;
 
@@ -145,6 +145,12 @@ public class VungleInterstitial extends CustomEventInterstitial {
         } else {
             MoPubLog.w(INTERSTITIAL_TAG + "Placement ID for this Ad Unit is not in serverExtras.");
             isAllDataValid = false;
+        }
+
+        if (serverExtras.containsKey(PLACEMENT_IDS_KEY)) {
+            MoPubLog.w(INTERSTITIAL_TAG + "No need to set placement IDs " +
+                    "in MoPub dashboard with Vungle SDK version " +
+                    com.vungle.warren.BuildConfig.VERSION_NAME);
         }
 
         return isAllDataValid;