+++ Vungle/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
+  * 6.3.24.0
+    * This version of the adapters has been certified with Vungle 6.3.24.
+    * Check if the passed Placement ID is a valid placement for the given App ID.
+
   * 6.3.17.1
     * Update the placement ID returned in the `getAdNetworkId` API (used to generate server-side rewarded video callback URL) to be non-null, and avoid potential NullPointerExceptions.